docker/Dockerfile: Create a docker-based way to build nvcomp
Building and then running nvcomp can be a bit tricky with dependencies
on different versions of cmake and CUDA etc. This commit creates a
Dockerfile which makes it much easier to build the nvcomp library and
benchmarks in a consistent environment. One can then either copy the
files out of the container onto the host or use the nvidia-docker [1]
framework to run programs against GPUs installed on the system.

[1]: https://github.com/NVIDIA/nvidia-docker

Addresses #52.